separate types from data
