+++ src/data/bundle-keys-types.ts
@@ -1,0 +1,1 @@
+export type BUNDLE = 'STUDENT' | 'INTERMEDIATE' | 'PROFESSIONAL';
