+++ src/data/bundle-keys.ts
@@ -31,5 +31,3 @@ export const THE_ROAD_TO_REACT_WITH_FIREBASE_BUNDLE_KEYS = {
   // RETIRED
   // BOOK: 'BOOK',
 };
-
-export type BUNDLE = 'STUDENT' | 'INTERMEDIATE' | 'PROFESSIONAL';
