+++ src/data/course-keys.ts
@@ -3,9 +3,3 @@ export const TAMING_THE_STATE = 'TAMING_THE_STATE';
 export const THE_ROAD_TO_GRAPHQL = 'THE_ROAD_TO_GRAPHQL';
 export const THE_ROAD_TO_REACT_WITH_FIREBASE =
   'THE_ROAD_TO_REACT_WITH_FIREBASE';
-
-export type COURSE =
-  | typeof THE_ROAD_TO_LEARN_REACT
-  | typeof TAMING_THE_STATE
-  | typeof THE_ROAD_TO_GRAPHQL
-  | typeof THE_ROAD_TO_REACT_WITH_FIREBASE;
